+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 - Added `NotCurrentContext::make_current_surfaceless(self)` and
   `PossiblyCurrentContext::make_current_surfaceless(&self)` in the `Wgl`
   implementation to allow the use of surfaceless contexts with WGL.
-- Added workaround for EGL drivers reporting KHR_platform_gbm with EGL 1.4.
+- Added workaround for EGL drivers reporting `EGL_KHR_platform_gbm` with EGL 1.4.
 
 
 # Version 0.32.1